Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,10 +4,10 @@
 
 int main() {
     int x,y,z ;
-    int Average;
+    float Average;
     scanf("%d", &x);
     scanf("%d", &y);
     scanf("%d", &z);
     Average = (x+y+z)/3.0;
-    printf("Average: %.2d", Average );
+    printf("Average: %.2f", Average );
 }
